It is better to use (ssize_t) instead of (int) when refering to string lengths.

+++ include/ircaux.h
@@ -111,7 +111,7 @@ double	time_to_next_minute 	(void);
 char *	remove_trailing_spaces 	(char *, size_t *cluep);
 char *	ltoa 			(long);
 char *	ftoa			(double);
-char *	strformat 		(char *, const char *, int, int);
+char *	strformat 		(char *, const char *, ssize_t, int);
 char *	chop_word 		(char *);
 char *	skip_spaces		(char *);
 int	split_args		(char *, char **to, size_t);
